Add option to create multiple read replica DBs in PaaS

+++ src/main/IaC/PaaS/destroy.sh
@@ -1,11 +1,23 @@
 #!/bin/bash
 
-echo "-----------------------------------------------------"
-echo "Destroying all objects in tested resource group"
-terraform destroy
-. ./variables.sh
-az group delete --name paas-spring-rg
+echo "Are you sure that you want to destroy the test environment? (y/n)"
+read -r answer
 
-echo "-----------------------------------------------------"
-echo "Restoring original files"
-mv terraform.tfvars.backup terraform.tfvars
+if [ "$answer" = "y" ] || [ "$answer" = "Y" ]; then
+    echo "-----------------------------------------------------"
+    echo "Destroying all objects in tested resource group"
+    terraform destroy --auto-approve
+    . ./variables.sh
+    az group delete --name paas-spring-rg --yes
+
+    echo "-----------------------------------------------------"
+    echo "Restoring original files"
+    mv terraform.tfvars.backup terraform.tfvars
+
+    echo "-----------------------------------------------------"
+    echo "The test environment has been cleaned"
+elif [ "$answer" = "n" ] || [ "$answer" = "N" ]; then
+    echo "Aborting. Test environment will not be destroyed."
+else
+    echo "Invalid input. Please enter 'y' or 'n'."
+fi
